moving pieces to right places
