+++ Cargo.toml
@@ -30,7 +30,7 @@ members = [
 [dependencies]
 clap = "3.0.0-beta.2"
 lmdb-rkv = "0.14.0"
-nom = "7"
+nom = "6.1.0"
 rayon = "1.5"
 thiserror = "1.0"
 codespan-reporting = "0.11"
