+++ crates/rules/src/lib.rs
@@ -16,6 +16,8 @@ pub use self::set::Set;
 pub use self::subject::Part as SubjPart;
 pub use self::subject::Subject;
 
+pub mod parser;
+
 mod decision;
 mod file_type;
 mod object;
