+++ crates/rules/src/parser/error.rs
@@ -1,0 +1,17 @@
+use nom::error::{ErrorKind, ParseError};
+
+#[derive(Debug, PartialEq)]
+pub enum CustomError<I> {
+    Generic,
+    Nom(I, ErrorKind),
+}
+
+impl<I> ParseError<I> for CustomError<I> {
+    fn from_error_kind(input: I, kind: ErrorKind) -> Self {
+        CustomError::Nom(input, kind)
+    }
+
+    fn append(_: I, _: ErrorKind, other: Self) -> Self {
+        other
+    }
+}
