+++ crates/rules/src/parser/mod.rs
@@ -1,0 +1,5 @@
+pub mod error;
+pub mod trace;
+
+#[cfg(test)]
+mod tests;
